requirements.txt
annotated-types==0.6.0
anyio==4.2.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.4
cryptography==41.0.7
Deprecated==1.2.14
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.109.0
firebase-admin==6.3.0
gcloud==0.18.3
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jwcrypto==1.5.1
MarkupSafe==2.1.3
msgpack==1.0.7
oauth2client==4.1.3
orjson==3.9.10
packaging==24.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
pytest==8.1.1
pytest-cov==5.0.0
python-dotenv==1.0.0
python-jwt==4.1.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.2.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
stripe==7.12.0
typing_extensions==4.9.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0